refactor: move express setup and bootstrap to main layer
