+++ customer/src/presentation/rest/route/index.ts
@@ -1,0 +1,1 @@
+export * from './customer.route'
